requirements.txt
streamlit==1.40.1
botorch==0.11.0
plotly==5.22.0
pandas
numpy